backend/requirements.txt
# Web framework
fastapi>=0.109.0,<0.120.0
uvicorn[standard]>=0.27.0,<0.30.0
python-multipart>=0.0.6,<0.1.0

# LangChain / LangGraph - use compatible versions
langgraph>=0.2.0,<0.3.0
langchain>=0.2.0,<0.3.0
langchain-openai>=0.1.0,<0.2.0

# OpenAI
openai>=1.10.0,<2.0.0

# Environment
python-dotenv>=1.0.0,<2.0.0

# Supabase
supabase>=2.0.0,<3.0.0

# PDF processing
PyPDF2>=3.0.0,<4.0.0
pymupdf>=1.24.0,<2.0.0
Pillow>=10.0.0,<11.0.0

# Data validation
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0

# Authentication
python-jose[cryptography]>=3.3.0,<4.0.0
bcrypt>=4.1.0,<5.0.0

# HTTP client
httpx>=0.24.0,<1.0.0

# Tokenization
tiktoken>=0.5.0,<1.0.0
